Change percentage from integer to float

+++ db/migrate/20221111144943_create_bulk_discounts.rb
@@ -1,7 +1,7 @@
 class CreateBulkDiscounts < ActiveRecord::Migration[5.2]
   def change
     create_table :bulk_discounts do |t|
-      t.integer :percentage
+      t.float :percentage
       t.integer :quantity_threshold
       t.references :merchant, foreign_key: true
       t.timestamps
